+++ commons/ui/src/main/res/values/colors.xml
@@ -13,5 +13,4 @@
     <!--Freaks screen -->
     <color name="card_view_color">#C8C8C8</color>
     <color name="toolbar_color">#808080</color>
-
 </resources>
