+++ features/freaks/build.gradle.kts
@@ -9,6 +9,12 @@ plugins {
 android {
     addProductFlavours(this)
 }
+
+dependencies {
+    implementation(project(BuildModules.Commons.UI))
+    implementation(project(BuildModules.Commons.VIEWS))
+}
+
 object PluginsVersions {
     const val LIFECYCLE_VIEW_MODEL = "2.3.1"
 }
