+++ features/freaks/src/main/kotlin/com/agilefreaks/freaks_catalog/features/freaks/DetailsViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
 import java.lang.Exception
 
 class DetailsViewModel(private val repository: FreakDetailsRepository) : ViewModel() {
