+++ features/freaks/src/main/kotlin/com/agilefreaks/freaks_catalog/features/freaks/Freak.kt
@@ -1,7 +1,14 @@
 package com.agilefreaks.freaks_catalog.features.freaks
 
-data class  Freak (
-    val id: String,
+data class Freak(
+    val freakId:String,
     val firstName: String,
-    val photo: String
+    val lastName: String,
+    val role: String,
+    val norm: String,
+    val level: String,
+    val description: String,
+    val photo: String,
+    val skills: String,
+    val projects: String
 )
