+++ features/freaks/src/main/kotlin/com/agilefreaks/freaks_catalog/features/freaks/FreakList.kt
@@ -1,0 +1,7 @@
+package com.agilefreaks.freaks_catalog.features.freaks
+
+data class  FreakList (
+    val id: String,
+    val firstName: String,
+    val photo: String
+)
