+++ features/freaks/src/main/res/values/strings.xml
@@ -4,7 +4,6 @@
     <string name="projects_btn">Projects</string>
     <string name="resetBtn">Reset</string>
     <string name="applyFiltersBtn">Apply</string>
-    <string name="resetBtn">Reset button</string>
     <string name="app_name">freaks-details-screen</string>
     <string name="description_template">
         Lorem ipsum dolor sit amet, consecte adipiscing elit. Aenean commodo eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis part montes, nascetur ridiculus mus.
