+++ features/freaks/src/test/kotlin/com/agilefreaks/freaks_catalog/features/freaks/model/DetailsViewModelTest.kt
@@ -1,0 +1,5 @@
+package com.agilefreaks.freaks_catalog.features.freaks.model
+
+import org.junit.Assert.*
+
+class DetailsViewModelTest
