+++ features/freaks/src/test/kotlin/com/agilefreaks/freaks_catalog/features/freaks/model/FreaksViewModelTest.kt
@@ -2,7 +2,7 @@ package com.agilefreaks.freaks_catalog.features.freaks.model
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.agilefreaks.freaks_catalog.features.freaks.CoroutineRule
-import com.agilefreaks.freaks_catalog.features.freaks.Freak
+import com.agilefreaks.freaks_catalog.features.freaks.FreakList
 import com.agilefreaks.freaks_catalog.features.freaks.repository.FreaksRepository
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -21,7 +21,7 @@ class FreaksViewModelTest {
     @Test
     fun `freaks will populate on init`() {
         // setup
-        val freak = Freak(
+        val freak = FreakList(
             id = "22",
             firstName = "Robert",
             photo = "https://i.ibb.co/kK7MQQD/rice.jpg"
@@ -34,10 +34,10 @@ class FreaksViewModelTest {
         assertThat(viewModel.freaks.value).containsExactly(freak)
     }
 
-    class FreaksRepositoryMock(private val freaks: MutableList<Freak> = mutableListOf()) :
+    class FreaksRepositoryMock(private val freakLists: MutableList<FreakList> = mutableListOf()) :
         FreaksRepository {
-        override suspend fun getFreaksFromApi(): List<Freak> = freaks
+        override suspend fun getFreaksFromApi(): List<FreakList> = freakLists
 
-        fun add(freak: Freak) = freaks.add(freak)
+        fun add(freakList: FreakList) = freakLists.add(freakList)
     }
 }
